requirement.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
